// frontend/src/types/RecipeMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface RecipeMetadata {
  id: number;
  name: string;
  description: string | null;
  image_url: string | null;
  servings: number | null;
  work_time: number | null;
  wait_time: number | null;
  author: string;
  created_at: number;
}